UI: Moving clock to header, station alerts to header, header play button to now-playing card header
